Rename Bluetooth active_update_coordinator to active_update_processor
This rename is inline with the passive_update_processor.

We are going to need an active_update_coordinator to match
similar functionality to passive_update_coordinator for passive
devices that need an active connection and we forgot to
rename this one when we created the processor coordinator
