+++ homeassistant/components/bluetooth/active_update_processor.py
